Moved makefont functions to makefont-lib and included it from makefont
